Split flag verification for cleaner boskos logic
Also: location refactor, typo, better defaulting of boskos
